WIP Refactoring search abstraction
With this refactoring it is easier to switch between backends (see SEARCH_ENGINE_BACKEND_CLASS). I've adapted tests to run after the refactoring but some things are still pending including the refactoring of QueryManager() to remove Solr concepts from non-search-backend code.
